Broke out the store into multiple files. Formatted code
