+++ src/pages/Home.js
@@ -1,12 +1,12 @@
 import React from 'react';
 import { connect } from 'react-redux';
-import { logout } from '../store';
+import { logout } from '../redux/store';
 
 const _Home = ({ auth, logout }) => (
   <div className="container">
     <h2>Home</h2>
     <h4>Welcome {auth.id ? auth.firstName : "Guest"}!</h4>
-      <button
+    <button
       type="button"
       className="btn btn-outline-secondary"
       onClick={logout}
