+++ src/redux/constants.js
@@ -1,0 +1,10 @@
+/////////////////////////ACTION TYPES//////////////////////////
+export const SET_AUTH = 'SET_AUTH';
+export const SET_PRODUCTS = 'SET_PRODUCTS';
+export const SET_USERS = 'SET_USERS';
+export const UPDATE_USER = 'UPDATE_USER';
+export const SET_ORDERS = 'SET_ORDERS';
+export const SET_ORDERPRODUCTS = 'SET_ORDERPRODUCTS';
+export const DELETE_ORDERPRODUCT = 'DELETE_ORDERPRODUCT';
+export const UPDATE_ORDERPRODUCT = 'UPDATE_ORDERPRODUCT';
+export const ADD_ORDERPRODUCT = 'ADD_ORDERPRODUCT';
